Strip AbstractSynapseIO
Similar reasoning, except this has *never* had alternate impls
