+++ spynnaker/pyNN/models/neuron/max_row_info.py
