+++ unittests/model_tests/neuron/test_synapse_io_row_based.py
@@ -21,7 +21,7 @@ from spynnaker.pyNN.models.neuron.synapse_dynamics import (
     SynapseDynamicsStatic, SynapseDynamicsSTDP)
 from spynnaker.pyNN.models.neuron.master_pop_table_as_binary_search import (
     MasterPopTableAsBinarySearch)
-from spynnaker.pyNN.models.neuron.synapse_io import SynapseIORowBased
+from spynnaker.pyNN.models.neuron.synapse_io_row_based import SynapseIORowBased
 from spynnaker.pyNN.models.neuron.plasticity.stdp.weight_dependence import (
     WeightDependenceAdditive)
 from spynnaker.pyNN.models.neuron.plasticity.stdp.timing_dependence import (
